Re arrange file structure
