--- data/processed/data_exemple.csv
@@ -1,19 +1,0 @@
-id_post,class,subject,text,Media_url
-500388199064420352,0,ferguson-all-rnr-threads,"Hard to believe this guy assaulted a cop. 😂😂😂 
-
-#Ferguson http://t.co/J86VLXnZF6",http://pbs.twimg.com/media/BvG8ak6IQAAE2zx.jpg
-500306187624407040,0,ferguson-all-rnr-threads,The day #Ferguson cops beat a man and charged him with getting his blood on their uniforms http://t.co/P2mW9Nu8pB … http://t.co/6kBUzLXzFY,http://pbs.twimg.com/media/BvFrRW0CUAIN9JE.jpg
-500413818368184321,0,ferguson-all-rnr-threads,"Remarkably, despite the military-style police presence in #Ferguson, a known killer was able to skip town http://t.co/lDGjp6Wt1j",
-500292426154528770,0,ferguson-all-rnr-threads,"If Mike Brown was engaged in a robbery before being shot, why isn't his alleged accomplice under arrest? #Ferguson http://t.co/upMLU4bbK4",
-500360889347096576,0,ferguson-all-rnr-threads,so ... they clearly released that video  only to shame &amp; blame the victim. #Ferguson #MikeBrown,
-500289307257868288,0,ferguson-all-rnr-threads,The selective release of information by #Ferguson PD tells us nothing about the shooting. Instead appears to be a basic smear job.,
-500309222635360256,0,ferguson-all-rnr-threads,"Every incident report I've seen says ""the suspect was X,"" ""the suspect did Y."" On the #Ferguson report, #MikeBrown is ALREADY NAMED. HOW??",
-500375190535421953,0,ferguson-all-rnr-threads,"If anyone's looking for a reason to murder me execution-style in the street, I stole a coke from a Walmart when I was 19. #Ferguson",
-499404849835175938,0,ferguson-all-rnr-threads,#Ferguson is completely shut down. No journalists allowed.,
-500366333461422080,0,ferguson-all-rnr-threads,#Ferguson police chief says officer Darren Wilson didn't know #MikeBrown was a robbery suspect http://t.co/SbpxzwwZWs,
-500279160795721728,0,ferguson-all-rnr-threads,"Question for #Ferguson police chief? If OFFICER DARREN WILSON was in pursuit of a robbery suspect, why tell kids to get out of the road? LIE",
-500283877567770624,0,ferguson-all-rnr-threads,#Ferguson police say Michael Brown fit description of strong-arm robbery suspect  http://t.co/bxXMNDzhe6,
-500258409988763649,0,ferguson-all-rnr-threads,"Name of #Ferguson cop is expected to be released by 9 ET, police chief says: http://t.co/NhMsTgfzYi (J.B. Forbes, AP) http://t.co/7LU1iT1EKB",http://pbs.twimg.com/media/BvFGX38IUAAzoNy.jpg
-500357713260724224,0,ferguson-all-rnr-threads,"ICYMI: The info given today seems to line up with this witness account, even the alleged theft: http://t.co/7hjBIrf1a3
-
-#Ferguson #MikeBrown",
