+++ src/models/yolo/__init__.py
@@ -1,0 +1,1 @@
+from .objects_extractor import ObjectExtractor
