--- src/utils/yolo_utils.py
@@ -1,8 +1,0 @@
-"""
--- TO-DO --
-
-Create a Class that use yolo with the following functions :
-
-* get_labels(id:int) -> List[str] ; returns a set of labels of detected object as a list of strings, for exampel ['gun','person','airplane','book']
-
-"""
